+++ modules/ccv2/core/src/sap/commerce/toolset/ccv2/dto/CCv2DeploymentStatus.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
- * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * Copyright (C) 2019-2026 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -21,7 +21,7 @@ package sap.commerce.toolset.ccv2.dto
 import sap.commerce.toolset.HybrisIcons
 import javax.swing.Icon
 
-enum class CCv2DeploymentStatusEnum(val title: String, val icon: Icon) {
+enum class CCv2DeploymentStatus(val title: String, val icon: Icon) {
     SCHEDULED("Scheduled", HybrisIcons.CCv2.Deployment.Status.SCHEDULED),
     DEPLOYED("Deployed", HybrisIcons.CCv2.Deployment.Status.DEPLOYED),
     DEPLOYING("Deploying", HybrisIcons.CCv2.Deployment.Status.DEPLOYING),