+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## [2026.0.2]
+
+<cite>Release contributors</cite>
+
+- 1 PR(s) by [Mykhailo Lytvyn](https://github.com/epam/sap-commerce-intellij-idea-plugin/pulls?q=milestone%3A2026.0.2+author%3Amlytvyn+is%3Apr)
+
+### `CCv2` enhancements
+- Display scheduled activities within environment details [#1753](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1753)
+
 ## [2026.0.1]
 
 <cite>Release contributors</cite>